Convert Think assistant example to framework conventions
Move the assistant example onto the Think framework shape without weakening its auth-gated architecture. The production Worker now uses virtual:think/entry as main, keeps src/server.ts as the GitHub OAuth and /chat gate, and delegates chat tails through the injected Think router so friendly subagent paths resolve through the generated manifest instead of exposing direct /agents routes.

Restructure assistant agent code under agents/assistant. AssistantDirectory remains a plain Agent container that owns per-user directory state, shared workspace, MCP registry, and subagent orchestration, while MyAssistant becomes a generated facet under agents/assistant/agents. Update imports, client route usage, tests, env types, and docs to reflect the generated class names and the deliberate absence of public /agents/* routing for the assistant app.

The agent-skills framework conversion is handled separately against the landed agents/skills API.
