+++ examples/assistant/env.d.ts
@@ -1,22 +1,16 @@
 /* eslint-disable */
-// Generated by Wrangler by running `wrangler types env.d.ts --include-runtime false` (hash: 55e4ce52fc087c1e748b18ae35f2bddd)
-declare namespace Cloudflare {
-  interface GlobalProps {
-    mainModule: typeof import("./src/server");
-    durableNamespaces: "AssistantDirectory" | "MyAssistant";
-  }
-  interface Env {
-    LOADER: WorkerLoader;
-    AI: Ai;
-    GITHUB_CLIENT_ID: string;
-    GITHUB_CLIENT_SECRET: string;
-    AssistantDirectory: DurableObjectNamespace<
-      import("./src/server").AssistantDirectory
-    >;
-    MyAssistant: DurableObjectNamespace<import("./src/server").MyAssistant>;
-  }
+// Generated by Wrangler by running `wrangler types env.d.ts --include-runtime false` (hash: 2b0c06b546741fc31b1d905640fbe299)
+interface __BaseEnv_Env {
+  LOADER: WorkerLoader;
+  AI: Ai;
+  GITHUB_CLIENT_ID: string;
+  GITHUB_CLIENT_SECRET: string;
+  AssistantDirectory: DurableObjectNamespace /* ThinkAgent_Assistant */;
 }
-interface Env extends Cloudflare.Env {}
+declare namespace Cloudflare {
+  interface Env extends __BaseEnv_Env {}
+}
+interface Env extends __BaseEnv_Env {}
 type StringifyValues<EnvType extends Record<string, unknown>> = {
   [Binding in keyof EnvType]: EnvType[Binding] extends string
     ? EnvType[Binding]
