+++ examples/assistant/vite.config.ts
@@ -1,9 +1,9 @@
 import { cloudflare } from "@cloudflare/vite-plugin";
 import tailwindcss from "@tailwindcss/vite";
 import react from "@vitejs/plugin-react";
-import agents from "agents/vite";
+import { think } from "@cloudflare/think/vite";
 import { defineConfig } from "vite";
 
 export default defineConfig({
-  plugins: [agents(), react(), cloudflare(), tailwindcss()]
+  plugins: [think(), react(), cloudflare(), tailwindcss()]
 });
